top_films_series/requirements.txt
altair==4.2.2
appdirs==1.4.4
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
attrs @ file:///work/ci_py311/attrs_1676822758931/work
Automat @ file:///tmp/build/80754af9/automat_1600298431173/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
bcrypt @ file:///work/ci_py311/bcrypt_1676825144644/work
blinker==1.6.2
brotlipy==0.7.0
cachetools==5.3.0
certifi @ file:///work/ci_py311/certifi_1676822443922/work/certifi
cffi @ file:///work/ci_py311/cffi_1676822533496/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
comm @ file:///work/ci_py311/comm_1677709131612/work
constantly==15.1.0
cryptography @ file:///work/ci_py311_2/cryptography_1679335572456/work
cssselect==1.1.0
debugpy @ file:///work/ci_py311/debugpy_1676824903649/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
dnspython==2.3.0
entrypoints==0.4
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
filelock @ file:///work/ci_py311/filelock_1676823024085/work
flit_core @ file:///croot/flit-core_1679397103445/work/source/flit_core
gitdb==4.0.10
GitPython==3.1.31
hyperlink @ file:///tmp/build/80754af9/hyperlink_1610130746837/work
idna @ file:///work/ci_py311/idna_1676822698822/work
importlib-metadata==6.5.0
incremental @ file:///tmp/build/80754af9/incremental_1636629750599/work
ipykernel @ file:///work/ci_py311/ipykernel_1677867938687/work
ipython @ file:///croot/ipython_1680701871216/work
itemadapter @ file:///tmp/build/80754af9/itemadapter_1626442940632/work
itemloaders @ file:///opt/conda/conda-bld/itemloaders_1646805235997/work
jedi @ file:///work/ci_py311_2/jedi_1679336495545/work
Jinja2==3.1.2
jmespath @ file:///Users/ktietz/demo/mc3/conda-bld/jmespath_1630583964805/work
jsonschema==4.17.3
jupyter_client @ file:///croot/jupyter_client_1680171862562/work
jupyter_core @ file:///croot/jupyter_core_1679906564508/work
lxml @ file:///croot/lxml_1679646450407/work
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib-inline @ file:///work/ci_py311/matplotlib-inline_1676823841154/work
mdurl==0.1.2
nest-asyncio @ file:///work/ci_py311/nest-asyncio_1676823382924/work
numpy==1.24.2
packaging @ file:///croot/packaging_1678965309396/work
pandas==1.5.3
parsel @ file:///work/ci_py311/parsel_1676905937880/work
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.5.0
platformdirs @ file:///work/ci_py311/platformdirs_1676823055788/work
prompt-toolkit @ file:///work/ci_py311/prompt-toolkit_1676824857154/work
Protego @ file:///tmp/build/80754af9/protego_1598657180827/work
protobuf==3.20.3
psutil @ file:///work/ci_py311_2/psutil_1679337388738/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyarrow==11.0.0
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydeck==0.8.1b0
PyDispatcher==2.0.5
Pygments==2.15.1
PyHamcrest @ file:///tmp/build/80754af9/pyhamcrest_1615748656804/work
pymango==0.1.1
pymongo==4.3.3
Pympler==1.0.1
pyOpenSSL @ file:///croot/pyopenssl_1678965284384/work
pyrsistent==0.19.3
PySocks @ file:///work/ci_py311/pysocks_1676822712504/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pyzmq @ file:///work/ci_py311/pyzmq_1676918885689/work
queuelib==1.5.0
requests @ file:///work/ci_py311_2/requests_1679337615230/work
requests-file @ file:///Users/ktietz/demo/mc3/conda-bld/requests-file_1629455781986/work
rich==13.3.4
Scrapy @ file:///work/ci_py311/scrapy_1677770267574/work
service-identity @ file:///Users/ktietz/demo/mc3/conda-bld/service_identity_1629460757137/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
streamlit==1.21.0
tldextract @ file:///opt/conda/conda-bld/tldextract_1646638314385/work
toml==0.10.2
toolz==0.12.0
tornado @ file:///work/ci_py311/tornado_1676823418767/work
traitlets @ file:///work/ci_py311/traitlets_1676823305040/work
Twisted @ file:///work/ci_py311/twisted_1676846057592/work
typing_extensions @ file:///work/ci_py311/typing_extensions_1676822515011/work
tzdata==2023.3
tzlocal==4.3
urllib3 @ file:///croot/urllib3_1680254681959/work
validators==0.20.0
w3lib @ file:///Users/ktietz/demo/mc3/conda-bld/w3lib_1629359764703/work
watchdog==3.0.0
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
zipp==3.15.0
zope.interface @ file:///work/ci_py311/zope.interface_1676905896080/work
